+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### next release
 - :bulb: Added feature to clean all contacts statuses
+- :bulb: Added hidden feature to mark contacts and friends randomly
 - :arrow_up: Refactored danger controller actions
 
 ### v0.6.5